#76 stergere amount + modificare migrare + service
