+++ backend/src/main/resources/db/migration/V1_2_0__create_payments_table.sql
@@ -6,7 +6,6 @@ CREATE TABLE payment (
                          name VARCHAR(255) NOT NULL,
                          status VARCHAR(10),
                          payment_date DATE,
-                         amount NUMERIC(19,2) NOT NULL,
                          CONSTRAINT fk_expense FOREIGN KEY(expense_id) REFERENCES expense(id)
 );
