+++ frontend/src/app/component/payments/add-payments/add-payments.component.css
@@ -162,4 +162,3 @@ body, .app-container {
   margin-top: 10px;
 }
 
-
