+++ frontend/src/app/component/payments/add-payments/add-payments.component.html
@@ -15,7 +15,7 @@
         <div class="form-field">
           <label for="name">Payment Name <span class="required">*</span></label>
           <input type="text" id="name" [(ngModel)]="payment.name" name="name"
-                 (input)="onNameInput()"
+                 (input)="onNameInput($event)"
                  [ngClass]="{'invalid-field': submitted && isNameInvalid}"
                  placeholder="Enter payment name">
           <p-message class="field-msg" [ngClass]="{'show': submitted && isNameInvalid}"
@@ -33,10 +33,9 @@
                   (ngModelChange)="onStatusChange()"
                   [ngClass]="{'invalid-field': submitted && isStatusInvalid}">
             <option value="" disabled>Select status</option>
-            <option value="PENDING">Pending</option>
-            <option value="PAID">Paid</option>
-            <option value="FAILED">Failed</option>
-
+            <option value="pending">Pending</option>
+            <option value="paid">Paid</option>
+            <option value="failed">Failed</option>
           </select>
           <p-message class="field-msg" [ngClass]="{'show': submitted && isStatusInvalid}"
                      icon="pi pi-exclamation-triangle"
@@ -48,9 +47,9 @@
 
         <!-- Payment Date -->
         <div class="form-field">
-          <label for="paymentDate">Payment Date <span class="required">*</span></label>
-          <input type="date" id="paymentDate" [(ngModel)]="payment.paymentDate" name="paymentDate"
-                 (input)="onDateInput()"
+          <label for="payment_date">Payment Date <span class="required">*</span></label>
+          <input type="date" id="payment_date" [(ngModel)]="payment.paymentDate" name="payment_date"
+                 (input)="onDateInput($event)"
                  [ngClass]="{'invalid-field': submitted && isDateInvalid}">
           <p-message class="field-msg" [ngClass]="{'show': submitted && isDateInvalid}"
                      icon="pi pi-exclamation-triangle"
@@ -62,9 +61,9 @@
 
         <!-- Expense ID -->
         <div class="form-field">
-          <label for="expenseId">Expense ID <span class="required">*</span></label>
-          <input type="text" id="expenseId"
-                 name="expenseId"
+          <label for="expense_id">Expense ID <span class="required">*</span></label>
+          <input type="text" id="expense_id"
+                 name="expense_id"
                  [(ngModel)]="payment.expenseId"
                  (input)="onExpenseIdInput($event)"
                  [ngClass]="{'invalid-field': submitted && isExpenseIdInvalid}"
