+++ frontend/src/app/component/payments/edit-payments/edit-payments.component.html
@@ -8,126 +8,90 @@
       <div class="welcome">
         <h2>Edit Payment</h2>
       </div>
-      <p class="payment-subtitle">Update the details of your payment</p>
+      <p class="payment-subtitle">Edit the details of your payment</p>
 
       <form class="payment-form" #form="ngForm" (ngSubmit)="savePayment(form)">
         <!-- Payment Name -->
         <div class="form-field">
           <label for="name">Payment Name <span class="required">*</span></label>
-          <input
-            type="text"
-            id="name"
-            [(ngModel)]="payment.name"
-            name="name"
-            (input)="onNameInput()"
-            [ngClass]="{ 'invalid-field': submitted && isNameInvalid }"
-            placeholder="Enter payment name"
-          />
-          <p-message
-            class="field-msg"
-            [ngClass]="{ show: submitted && isNameInvalid }"
-            icon="pi pi-exclamation-triangle"
-            variant="simple"
-            severity="error"
-            text="Please provide a name for the payment."
-            size="small"
-          ></p-message>
+          <input type="text" id="name" [(ngModel)]="payment.name" name="name"
+                 (input)="onNameInput($event)"
+                 [ngClass]="{'invalid-field': submitted && isNameInvalid}"
+                 placeholder="Enter payment name">
+          <p-message class="field-msg" [ngClass]="{'show': submitted && isNameInvalid}"
+                     icon="pi pi-exclamation-triangle"
+                     variant="simple"
+                     severity="error"
+                     text="Please provide a name for the payment."
+                     size="small"></p-message>
         </div>
 
         <!-- Status -->
         <div class="form-field">
           <label for="status">Status <span class="required">*</span></label>
-          <select
-            id="status"
-            [(ngModel)]="payment.status"
-            name="status"
-            (ngModelChange)="onStatusChange()"
-            [ngClass]="{ 'invalid-field': submitted && isStatusInvalid }"
-          >
+          <select id="status" [(ngModel)]="payment.status" name="status"
+                  (ngModelChange)="onStatusChange()"
+                  [ngClass]="{'invalid-field': submitted && isStatusInvalid}">
             <option value="" disabled>Select status</option>
-            <option value="PENDING">Pending</option>
-            <option value="PAID">Paid</option>
-            <option value="FAILED">Failed</option>
+            <option value="pending">Pending</option>
+            <option value="paid">Paid</option>
+            <option value="failed">Failed</option>
           </select>
-          <p-message
-            class="field-msg"
-            [ngClass]="{ show: submitted && isStatusInvalid }"
-            icon="pi pi-exclamation-triangle"
-            variant="simple"
-            severity="error"
-            text="Please select a status for the payment."
-            size="small"
-          ></p-message>
+          <p-message class="field-msg" [ngClass]="{'show': submitted && isStatusInvalid}"
+                     icon="pi pi-exclamation-triangle"
+                     variant="simple"
+                     severity="error"
+                     text="Please select a status for the payment."
+                     size="small"></p-message>
         </div>
 
         <!-- Payment Date -->
         <div class="form-field">
-          <label for="paymentDate">Payment Date <span class="required">*</span></label>
-          <input
-            type="date"
-            id="paymentDate"
-            [(ngModel)]="payment.paymentDate"
-            name="paymentDate"
-            (input)="onDateInput()"
-            [ngClass]="{ 'invalid-field': submitted && isDateInvalid }"
-          />
-          <p-message
-            class="field-msg"
-            [ngClass]="{ show: submitted && isDateInvalid }"
-            icon="pi pi-exclamation-triangle"
-            variant="simple"
-            severity="error"
-            text="Please select a valid date for the payment."
-            size="small"
-          ></p-message>
+          <label for="payment_date">Payment Date <span class="required">*</span></label>
+          <input type="date" id="payment_date" [(ngModel)]="payment.paymentDate" name="payment_date"
+                 (input)="onDateInput($event)"
+                 [ngClass]="{'invalid-field': submitted && isDateInvalid}">
+          <p-message class="field-msg" [ngClass]="{'show': submitted && isDateInvalid}"
+                     icon="pi pi-exclamation-triangle"
+                     variant="simple"
+                     severity="error"
+                     text="Please select a valid date for the payment."
+                     size="small"></p-message>
         </div>
 
         <!-- Expense ID -->
         <div class="form-field">
-          <label for="expenseId">Expense ID <span class="required">*</span></label>
-          <input
-            type="text"
-            id="expenseId"
-            name="expenseId"
-            [(ngModel)]="payment.expenseId"
-            (input)="onExpenseIdInput($event)"
-            [ngClass]="{ 'invalid-field': submitted && isExpenseIdInvalid }"
-            placeholder="Enter numeric expense id"
-          />
-          <p-message
-            class="field-msg"
-            [ngClass]="{ show: submitted && isExpenseIdInvalid }"
-            icon="pi pi-exclamation-triangle"
-            variant="simple"
-            severity="error"
-            [text]="expenseErrorMessage"
-            size="small"
-          ></p-message>
+          <label for="expense_id">Expense ID <span class="required">*</span></label>
+          <input type="text" id="expense_id"
+                 name="expense_id"
+                 [(ngModel)]="payment.expenseId"
+                 (input)="onExpenseIdInput($event)"
+                 [ngClass]="{'invalid-field': submitted && isExpenseIdInvalid}"
+                 placeholder="Enter numeric expense id">
+          <p-message class="field-msg" [ngClass]="{'show': submitted && isExpenseIdInvalid}"
+                     icon="pi pi-exclamation-triangle"
+                     variant="simple"
+                     severity="error"
+                     [text]="expenseErrorMessage"
+                     size="small"></p-message>
         </div>
 
-        <!-- Buttons -->
         <div class="button">
-          <p-toast
-            #toast
-            position="top-left"
-            style="margin-left: 600px; margin-bottom: 140px"
+          <!-- Edit Button -->
+          <p-toast #toast position="top-left"
+                 style="margin-left: 600px; margin-bottom: 140px;"
           ></p-toast>
-
-          <button
-            type="button"
-            pRipple
-            class="submit-btn"
-            (click)="savePayment(form)"
-          >
+          <button type="button" pRipple class="submit-btn" (click)="savePayment(form)">
             Edit Payment
           </button>
 
+          <!-- Cancel Button -->
           <button type="button" class="submit-btn" (click)="cancelEdit()">
-            Cancel
+             Cancel
           </button>
         </div>
 
-        <!-- Info Section -->
+        <!-- Keep Track Section -->
         <div class="piggy-bank-section">
           <div class="piggy-bank-icon">
             <i class="pi pi-sparkles"></i>
