+++ frontend/src/app/component/payments/list-payments/all-payments.component.css
@@ -151,70 +151,3 @@ body, .app-container {
   align-items: center;
   gap: 6px;
 }
-/* ===== PrimeNG ConfirmDialog Custom Theme ===== */
-::ng-deep .p-confirm-dialog {
-  border-radius: 14px !important;
-  overflow: hidden !important;
-  box-shadow: 0 8px 25px rgba(0,0,0,0.25) !important;
-}
-
-::ng-deep .p-confirm-dialog .p-dialog-header {
-  background: linear-gradient(to right, #A16EFF, #6A0DAD) !important;
-  color: white !important;
-  font-weight: 600 !important;
-  font-family: 'Poppins', sans-serif !important;
-  font-size: 16px !important;
-  border: none !important;
-}
-
-::ng-deep .p-confirm-dialog .p-dialog-content {
-  background-color: #F9F6FC !important;
-  color: #4B0082 !important;
-  font-family: 'Poppins', sans-serif !important;
-  font-size: 14px !important;
-  padding: 20px 25px !important;
-}
-
-::ng-deep .p-confirm-dialog .p-dialog-footer {
-  display: flex !important;
-  flex-direction: row-reverse !important;
-  justify-content: flex-end !important;
-  gap: 12px !important;
-}
-
-
-::ng-deep .p-confirm-dialog .p-button-success {
-  background: linear-gradient(to right, #7EE081, #47B04B) !important;
-  border: none !important;
-  color: white !important;
-  font-weight: 600;
-  font-family: 'Poppins', sans-serif;
-  padding: 6px 18px;
-  border-radius: 6px;
-  transition: 0.2s ease-in-out;
-}
-
-::ng-deep .p-confirm-dialog .p-button-success:hover {
-  background: linear-gradient(to right, #47B04B, #7EE081) !important;
-  transform: scale(1.03);
-}
-
-
-::ng-deep .p-confirm-dialog .p-button-text {
-  color: #6A0DAD !important;
-  font-weight: 600 !important;
-  font-family: 'Poppins', sans-serif;
-  transition: 0.2s ease-in-out;
-}
-
-::ng-deep .p-confirm-dialog .p-button-text:hover {
-  text-decoration: underline;
-  color: #A16EFF !important;
-}
-
-::ng-deep .p-toast .p-toast-message-success .p-toast-detail {
-  color: black !important;
-  font-family: 'Poppins', sans-serif;
-  font-weight: 500;
-}
-
