+++ frontend/src/app/component/payments/overview-payments/overview.component.html
@@ -13,30 +13,18 @@
       <!-- Dashboard Cards -->
       <div class="cards">
         <p-card class="card total-payments">
-          <i class="pi pi-wallet"></i>
-          <div>
-            <h3>Total Payments</h3>
-            <p>{{ totalPayments }}</p>
-          </div>
+          <i class="pi pi-money-bill"></i>
+          <div><h3>Total Payments</h3><p>$1,850</p></div>
         </p-card>
-
-        <p-card class="card paid-payments">
-          <i class="pi pi-check-circle"></i>
-          <div>
-            <h3>Paid</h3>
-            <p>{{ paidPayments }}</p>
-          </div>
+        <p-card class="card card-payments">
+          <i class="pi pi-credit-card"></i>
+          <div><h3>Card</h3><p>$1,100</p></div>
         </p-card>
-
-        <p-card class="card pending-payments">
-          <i class="pi pi-clock"></i>
-          <div>
-            <h3>Pending</h3>
-            <p>{{ pendingPayments }}</p>
-          </div>
+        <p-card class="card transfer-payments">
+          <i class="pi pi-arrow-right"></i>
+          <div><h3>Transfer</h3><p>$750</p></div>
         </p-card>
       </div>
-
 
       <!-- Chart -->
       <p-card class="chart-card">
@@ -49,16 +37,31 @@
         <!-- Recent Transactions -->
         <p-card class="long-card recent-activity">
           <h4>Recent Payment Transactions</h4>
-          <div *ngFor="let tx of transactions; let i = index" class="activity-item">
-            <i class="pi pi-money-bill"></i>
-            <p>{{ tx.name }} <span>{{ tx.paymentDate }}</span></p>
+          <div *ngFor="let tx of transactions; let i = index"
+               [ngClass]="{'card-payments': tx.type==='card','transfer-payments': tx.type==='transfer'}"
+               class="activity-item">
+
+            <i [class]="tx.icon"></i>
+            <p>{{tx.name}} <span>{{tx.date}}</span></p>
 
             <div>
-              <button pButton type="button" icon="pi pi-pencil" label="Edit" class="button edit"
-                      (click)="editPayment(i)"></button>
+              <!-- Edit -->
+              <button pButton
+                      type="button"
+                      icon="pi pi-pencil"
+                      label="Edit"
+                      class="button edit"
+                      (click)="editPayment(i)">
+              </button>
 
-              <button pButton type="button" icon="pi pi-trash" label="Delete" class="button delete"
-                      (click)="confirmDelete(i)"></button>
+              <!-- Delete -->
+              <button pButton
+                      type="button"
+                      icon="pi pi-trash"
+                      label="Delete"
+                      class="button delete"
+                      (click)="deletePayment(i)">
+              </button>
             </div>
           </div>
 
@@ -75,8 +78,5 @@
         </p-card>
       </div>
     </div>
-    <p-confirmDialog></p-confirmDialog>
-    <p-toast  position="top-left"
-              style="margin-left: 600px; margin-bottom: 140px;"></p-toast>
   </div>
 </div>
