--- frontend/src/app/payments/all/all-payments.component.css
@@ -1,153 +1,0 @@
-/* ===== Body & Container ===== */
-body, .app-container {
-  margin: 0;
-  padding: 0;
-  font-family: 'Poppins', sans-serif;
-  height: 100vh;
-  display: flex;
-  background-color: #F0E6FA;
-  overflow: hidden;
-}
-
-/* ===== Content ===== */
-.content {
-  flex: 1;
-  background: #fff;
-  border-top-left-radius: 16px;
-  border-bottom-left-radius: 16px;
-  box-shadow: -4px 0 10px rgba(0,0,0,0.05);
-  overflow-y: auto;
-  display: flex;
-  justify-content: center;
-}
-
-.page-wrapper {
-  width: 100%;
-  max-width: 1120px;
-  padding: 25px;
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
-}
-
-/* ===== Header + Add Payment ===== */
-.payment-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.payment-controls {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  gap: 10px;
-}
-
-.filters input, .filters select {
-  background-color: #F9F6FC;
-  color: #4B0082;
-  border: 1px solid #DCC6E0;
-  padding: 6px 10px;
-  border-radius: 6px;
-  margin-right: 8px;
-  font-family: 'Poppins', sans-serif;
-}
-
-#addPaymentBtn {
-  padding: 12px 24px;
-  background: linear-gradient(to right,#A16EFF,#6A0DAD);
-  border: none;
-  border-radius: 6px;
-  color: white;
-  font-weight: bold;
-  cursor: pointer;
-  font-family: 'Poppins', sans-serif;
-}
-
-#addPaymentBtn:hover {
-  background: linear-gradient(to right,#6A0DAD,#A16EFF);
-}
-
-/* ===== Payments Table ===== */
-.payment-table {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  width: 100%;
-  font-family: 'Poppins', sans-serif;
-}
-
-.table-header, .table-row {
-  display: grid;
-  grid-template-columns: 0.7fr 1.2fr 1.0fr 1.2fr 0.9fr;
-  padding: 14px;
-  border-radius: 8px;
-  align-items: center;
-  gap: 12px;
-  color: #4B0082;
-}
-
-.table-header {
-  background: #E6CCF2;
-  font-weight: bold;
-  color: #4B0082;
-}
-
-.table-row {
-  background: #F9F6FC;
-  box-shadow: 0 2px 5px rgba(0,0,0,0.05);
-}
-
-.table-row:nth-child(even) {
-  background: #F3E6FA;
-}
-
-.table-row span {
-  display: block;
-  padding: 4px 6px;
-}
-.table-row span:last-child, .table-header span:last-child {
-  display: flex;
-  justify-content: center;
-  gap: 8px;
-}
-
-/* ===== Buttons ===== */
-.button {
-  font-family: 'Poppins', sans-serif;
-  font-size: 12px;
-  border-radius: 6px;
-  padding: 5px 10px;
-  margin-left: 4px;
-}
-
-.button.edit {
-  background: linear-gradient(to right, #A16EFF, #C080FF);
-
-}
-.pi-pencil, .pi-trash{
-  font-size: 11px;
-}
-.button.edit:hover {
-  background: linear-gradient(to right, #C080FF, #A16EFF);
-}
-
-.button.delete {
-  background: linear-gradient(to right, #FFB6D9, #FF99CC);
-
-}
-.button.delete:hover {
-  background: linear-gradient(to right, #FFB6D9, #FF99CC)
-}
-
-.button.delete, button.edit{
-  min-width: 75px;
-  color: white;
-  border: none;
-  font-size: 13px;
-  display: flex;
-  align-items: center;
-  gap: 6px;
-}
