--- frontend/src/app/payments/edit/edit-payments.component.html
@@ -1,106 +1,0 @@
-<div class="app-container">
-  <!-- Sidebar -->
-  <app-sidebar (menuSelect)="onMenuSelect($event)"></app-sidebar>
-
-  <!-- Content -->
-  <div class="content">
-    <div class="page-wrapper">
-      <div class="welcome">
-        <h2>Edit Payment</h2>
-      </div>
-      <p class="payment-subtitle">Edit the details of your payment</p>
-
-      <form class="payment-form" #form="ngForm" (ngSubmit)="savePayment(form)">
-        <!-- Payment Name -->
-        <div class="form-field">
-          <label for="name">Payment Name <span class="required">*</span></label>
-          <input type="text" id="name" [(ngModel)]="payment.name" name="name"
-                 (input)="onNameInput($event)"
-                 [ngClass]="{'invalid-field': submitted && isNameInvalid}"
-                 placeholder="Enter payment name">
-          <p-message class="field-msg" [ngClass]="{'show': submitted && isNameInvalid}"
-                     icon="pi pi-exclamation-triangle"
-                     variant="simple"
-                     severity="error"
-                     text="Please provide a name for the payment."
-                     size="small"></p-message>
-        </div>
-
-        <!-- Status -->
-        <div class="form-field">
-          <label for="status">Status <span class="required">*</span></label>
-          <select id="status" [(ngModel)]="payment.status" name="status"
-                  (ngModelChange)="onStatusChange()"
-                  [ngClass]="{'invalid-field': submitted && isStatusInvalid}">
-            <option value="" disabled>Select status</option>
-            <option value="pending">Pending</option>
-            <option value="paid">Paid</option>
-            <option value="failed">Failed</option>
-          </select>
-          <p-message class="field-msg" [ngClass]="{'show': submitted && isStatusInvalid}"
-                     icon="pi pi-exclamation-triangle"
-                     variant="simple"
-                     severity="error"
-                     text="Please select a status for the payment."
-                     size="small"></p-message>
-        </div>
-
-        <!-- Payment Date -->
-        <div class="form-field">
-          <label for="payment_date">Payment Date <span class="required">*</span></label>
-          <input type="date" id="payment_date" [(ngModel)]="payment.payment_date" name="payment_date"
-                 (input)="onDateInput($event)"
-                 [ngClass]="{'invalid-field': submitted && isDateInvalid}">
-          <p-message class="field-msg" [ngClass]="{'show': submitted && isDateInvalid}"
-                     icon="pi pi-exclamation-triangle"
-                     variant="simple"
-                     severity="error"
-                     text="Please select a valid date for the payment."
-                     size="small"></p-message>
-        </div>
-
-        <!-- Expense ID -->
-        <div class="form-field">
-          <label for="expense_id">Expense ID <span class="required">*</span></label>
-          <input type="text" id="expense_id"
-                 name="expense_id"
-                 [(ngModel)]="payment.expense_id"
-                 (input)="onExpenseIdInput($event)"
-                 [ngClass]="{'invalid-field': submitted && isExpenseIdInvalid}"
-                 placeholder="Enter numeric expense id">
-          <p-message class="field-msg" [ngClass]="{'show': submitted && isExpenseIdInvalid}"
-                     icon="pi pi-exclamation-triangle"
-                     variant="simple"
-                     severity="error"
-                     [text]="expenseErrorMessage"
-                     size="small"></p-message>
-        </div>
-
-        <div class="button">
-          <!-- Edit Button -->
-          <p-toast #toast position="top-left"
-                 style="margin-left: 600px; margin-bottom: 140px;"
-          ></p-toast>
-          <button type="button" pRipple class="submit-btn" (click)="savePayment(form)">
-            Edit Payment
-          </button>
-
-          <!-- Cancel Button -->
-          <button type="button" class="submit-btn" (click)="cancelEdit()">
-             Cancel
-          </button>
-        </div>
-
-        <!-- Keep Track Section -->
-        <div class="piggy-bank-section">
-          <div class="piggy-bank-icon">
-            <i class="pi pi-sparkles"></i>
-          </div>
-          <p class="text">
-            Make sure all details are correct before saving your changes. ✏️
-          </p>
-        </div>
-      </form>
-    </div>
-  </div>
-</div>
