--- frontend/src/app/payments/overview/overview.component.html
@@ -1,82 +1,0 @@
-<div class="app-container">
-  <!-- Sidebar -->
-  <app-sidebar (menuSelect)="onMenuSelect($event)"></app-sidebar>
-
-  <!-- Content -->
-  <div class="content">
-    <div class="page-wrapper">
-      <div class="welcome-add-container">
-        <h2>Payment Overview</h2>
-      </div>
-      <p class="subtitle">Here’s a snapshot of your payments this month</p>
-
-      <!-- Dashboard Cards -->
-      <div class="cards">
-        <p-card class="card total-payments">
-          <i class="pi pi-money-bill"></i>
-          <div><h3>Total Payments</h3><p>$1,850</p></div>
-        </p-card>
-        <p-card class="card card-payments">
-          <i class="pi pi-credit-card"></i>
-          <div><h3>Card</h3><p>$1,100</p></div>
-        </p-card>
-        <p-card class="card transfer-payments">
-          <i class="pi pi-arrow-right"></i>
-          <div><h3>Transfer</h3><p>$750</p></div>
-        </p-card>
-      </div>
-
-      <!-- Chart -->
-      <p-card class="chart-card">
-        <h4>Monthly Payments</h4>
-        <p-chart type="bar" [data]="paymentChartData" [options]="chartOptions"></p-chart>
-      </p-card>
-
-      <!-- Bottom Section -->
-      <div class="bottom-section">
-        <!-- Recent Transactions -->
-        <p-card class="long-card recent-activity">
-          <h4>Recent Payment Transactions</h4>
-          <div *ngFor="let tx of transactions; let i = index"
-               [ngClass]="{'card-payments': tx.type==='card','transfer-payments': tx.type==='transfer'}"
-               class="activity-item">
-
-            <i [class]="tx.icon"></i>
-            <p>{{tx.name}} <span>{{tx.date}}</span></p>
-
-            <div>
-              <!-- Edit -->
-              <button pButton
-                      type="button"
-                      icon="pi pi-pencil"
-                      label="Edit"
-                      class="button edit"
-                      (click)="editPayment(i)">
-              </button>
-
-              <!-- Delete -->
-              <button pButton
-                      type="button"
-                      icon="pi pi-trash"
-                      label="Delete"
-                      class="button delete"
-                      (click)="deletePayment(i)">
-              </button>
-            </div>
-          </div>
-
-          <button pButton type="button" label="See All" class="button see-all" (click)="goToAllPayments()"></button>
-        </p-card>
-
-
-        <!-- Tips -->
-        <p-card class="long-card tips">
-          <h4>Payment Tips</h4>
-          <ul>
-            <li *ngFor="let tip of tips">{{tip.title}} <span>{{tip.note}}</span></li>
-          </ul>
-        </p-card>
-      </div>
-    </div>
-  </div>
-</div>
