--- frontend/src/app/payments/sidebar/sidebar.component.css
@@ -1,72 +1,0 @@
-.sidebar {
-  flex: 0 0 280px;
-  max-width: 280px;
-  min-width: 240px;
-  background-color: #E6CCF2;
-  color: #4B0082;
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  box-shadow: 3px 0 10px rgba(0,0,0,0.1);
-}
-
-.logo {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 20px;
-}
-
-.logo-shape {
-  width: 60px;
-  height: 60px;
-  background: linear-gradient(135deg, #A16EFF, #6A0DAD);
-  border-radius: 16px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.logo-shape i { color: white; font-size: 24px; }
-
-.logo h1 {
-  font-family: 'Poppins', sans-serif;
-  font-weight: 600;
-  color: #4B0082;
-  margin-top: 8px;
-  text-align: center;
-  font-size: 25px;
-}
-
-.menu {
-  list-style: none;
-  padding: 0;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 18px;
-  flex-grow: 1;
-}
-
-.menu li {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  padding: 10px 12px;
-  border-radius: 8px;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  font-weight: 500;
-}
-
-.menu li.active, .menu li:hover {
-  background-color: #DCC6E0;
-  transform: translateX(5px);
-}
-
-.menu li.logout {
-  margin-top: auto;
-  border-top: 1px solid rgba(0,0,0,0.1);
-  padding-top: 12px;
-}
